+++ CHANGELOG.md
@@ -9,6 +9,7 @@ principles.
 -   Upgraded `thinktomorrow\url` dependency to `3.0.0`.
 -   Fixed: multiple select filter options were not all used as filter in the query, only the first one.
 -   Added: `src/Forms/Layouts/Card` HasCollapsible trait to toggle content visibility.
+-   Added: support for custom locale segment in the chief admin toast.
 
 ## 0.8.17 - 2024-04-10